+++ src/main/java/com/umc/product/notification/adapter/in/web/FcmController.java
@@ -3,6 +3,7 @@ package com.umc.product.notification.adapter.in.web;
 import com.umc.product.notification.adapter.in.web.dto.request.FcmRegistrationRequest;
 import com.umc.product.notification.adapter.in.web.dto.request.FcmTestSendRequest;
 import com.umc.product.notification.application.port.in.ManageFcmUseCase;
+import com.umc.product.notification.application.port.in.RefreshFcmTokenUseCase;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class FcmController implements FcmControllerApi {
 
     private final ManageFcmUseCase manageFcmUseCase;
+    private final RefreshFcmTokenUseCase refreshFcmTokenUseCase;
 
     @Override
     @PostMapping("/{memberId}")
@@ -25,7 +27,7 @@ public class FcmController implements FcmControllerApi {
             // TODO: 인증 적용 시 @PathVariable -> @AuthenticationPrincipal 변경 필요
             @PathVariable("memberId") Long userId,
             @RequestBody FcmRegistrationRequest request) {
-        manageFcmUseCase.refreshTokenAndSubscriptions(userId, request);
+        refreshFcmTokenUseCase.refreshTokenAndSubscriptions(userId, request);
     }
 
     @Override
