+++ src/main/java/com/umc/product/notification/application/port/in/ManageFcmUseCase.java
@@ -17,6 +17,4 @@ public interface ManageFcmUseCase {
 
     void sendMessageByTopic(TopicNotificationCommand command);
 
-    void refreshTokenAndSubscriptions(Long userId, FcmRegistrationRequest request);
-
 }
