+++ src/test/java/com/umc/product/notification/application/service/FcmServiceTest.java
@@ -4,12 +4,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
-import static org.mockito.BDDMockito.willDoNothing;
-import static org.mockito.BDDMockito.willThrow;
-import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
 
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.FirebaseMessagingException;
@@ -18,20 +14,17 @@ import com.google.firebase.messaging.TopicManagementResponse;
 import com.umc.product.member.application.port.out.LoadMemberPort;
 import com.umc.product.member.domain.Member;
 import com.umc.product.notification.adapter.in.web.dto.request.FcmRegistrationRequest;
-import com.umc.product.notification.application.port.in.ManageFcmTopicUseCase;
 import com.umc.product.notification.application.port.in.dto.NotificationCommand;
 import com.umc.product.notification.application.port.in.dto.TopicNotificationCommand;
 import com.umc.product.notification.application.port.out.LoadFcmPort;
 import com.umc.product.notification.application.port.out.SaveFcmPort;
 import com.umc.product.notification.domain.FcmToken;
 import com.umc.product.notification.domain.exception.FcmDomainException;
-import com.umc.product.notification.domain.exception.FcmErrorCode;
 import com.umc.product.global.exception.BusinessException;
 import java.lang.reflect.Constructor;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.Nested;
-import org.mockito.InOrder;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -46,7 +39,6 @@ class FcmServiceTest {
     @Mock LoadMemberPort loadMemberPort;
     @Mock LoadFcmPort loadFcmPort;
     @Mock SaveFcmPort saveFcmPort;
-    @Mock ManageFcmTopicUseCase manageFcmTopicUseCase;
 
     @InjectMocks FcmService sut;
 
@@ -264,75 +256,6 @@ class FcmServiceTest {
             assertThatThrownBy(() ->
                     sut.sendMessageByToken(new NotificationCommand(999L, "제목", "내용")))
                     .isInstanceOf(BusinessException.class);
-        }
-    }
-
-    @Nested
-    class refreshTokenAndSubscriptions {
-
-        @Test
-        void 구독해제_토큰갱신_재구독_순서로_정상_수행된다() {
-            // given
-            Member member = createMember(1L);
-            FcmToken existingToken = createFcmToken("old-token");
-            given(loadFcmPort.findByMemberId(1L)).willReturn(existingToken);
-            given(loadMemberPort.findById(1L)).willReturn(Optional.of(member));
-            willDoNothing().given(manageFcmTopicUseCase).unsubscribeAllTopicsByMemberId(1L);
-            willDoNothing().given(manageFcmTopicUseCase).subscribeAllTopicsByMemberId(1L);
-
-            // when
-            sut.refreshTokenAndSubscriptions(1L, new FcmRegistrationRequest("new-token"));
-
-            // then
-            InOrder inOrder = inOrder(manageFcmTopicUseCase, loadMemberPort);
-            inOrder.verify(manageFcmTopicUseCase).unsubscribeAllTopicsByMemberId(1L);
-            inOrder.verify(loadMemberPort).findById(1L); // registerFcmToken 내부 호출
-            inOrder.verify(manageFcmTopicUseCase).subscribeAllTopicsByMemberId(1L);
-        }
-
-        @Test
-        void 구독해제_실패해도_토큰갱신과_재구독이_진행된다() {
-            // given
-            Member member = createMember(1L);
-            FcmToken existingToken = createFcmToken("old-token");
-            given(loadFcmPort.findByMemberId(1L)).willReturn(existingToken);
-            given(loadMemberPort.findById(1L)).willReturn(Optional.of(member));
-            willThrow(new FcmDomainException(FcmErrorCode.TOPIC_UNSUBSCRIBE_FAILED))
-                    .given(manageFcmTopicUseCase).unsubscribeAllTopicsByMemberId(1L);
-            willDoNothing().given(manageFcmTopicUseCase).subscribeAllTopicsByMemberId(1L);
-
-            // when
-            sut.refreshTokenAndSubscriptions(1L, new FcmRegistrationRequest("new-token"));
-
-            // then: 토큰 갱신(registerFcmToken)과 재구독이 정상 수행됨
-            then(loadMemberPort).should().findById(1L);
-            then(manageFcmTopicUseCase).should().subscribeAllTopicsByMemberId(1L);
-        }
-
-        @Test
-        void 재구독_실패시_보상로직으로_이전_토큰_복구_및_재구독을_시도한다() {
-            // given
-            Member member = createMember(1L);
-            FcmToken existingToken = createFcmToken("old-token");
-            // 첫 번째 호출: refreshTokenAndSubscriptions 시작 시 기존 토큰 조회
-            // 두 번째 호출: registerFcmToken 내부에서 조회
-            // 세 번째 호출: 보상 로직에서 토큰 복구를 위한 조회
-            given(loadFcmPort.findByMemberId(1L)).willReturn(existingToken);
-            given(loadMemberPort.findById(1L)).willReturn(Optional.of(member));
-            willDoNothing().given(manageFcmTopicUseCase).unsubscribeAllTopicsByMemberId(1L);
-            willThrow(new FcmDomainException(FcmErrorCode.TOPIC_SUBSCRIBE_FAILED))
-                    .willDoNothing()
-                    .given(manageFcmTopicUseCase).subscribeAllTopicsByMemberId(1L);
-
-            // when & then
-            assertThatThrownBy(() ->
-                    sut.refreshTokenAndSubscriptions(1L, new FcmRegistrationRequest("new-token")))
-                    .isInstanceOf(FcmDomainException.class);
-
-            // 보상 로직: 토큰 복구를 위해 loadFcmPort.findByMemberId가 추가 호출됨
-            then(loadFcmPort).should(times(3)).findByMemberId(1L);
-            // 보상 로직: 재구독 재시도
-            then(manageFcmTopicUseCase).should(times(2)).subscribeAllTopicsByMemberId(1L);
         }
     }
 }
